+++ packages/core/src/modal-suggest/ModalSuggestControlState.ts
@@ -1,0 +1,5 @@
+export default interface ModalSuggestControlState {
+  show: boolean
+  focused: boolean
+  hovered: boolean
+}