+++ packages/mobile/src/content-suggest/ContentSuggestOptionModel.ts
@@ -1,0 +1,7 @@
+import {SuggestOptionModel} from '@qiwi/pijma-core'
+
+export default interface ContentSuggestOptionModel<V> extends SuggestOptionModel<V> {
+  title: string
+  logo: string
+  description: string
+}